:sparkles: [User Settings] Fix phone bug + uniformize user form
