+++ src/ControllerV2/UserController.php
@@ -5,7 +5,7 @@ namespace App\ControllerV2;
 use App\Entity\Centre;
 use App\Entity\User;
 use App\FormV2\ChangePasswordFormType;
-use App\FormV2\UserSettingsType;
+use App\FormV2\UserType;
 use App\ManagerV2\RelayManager;
 use App\ManagerV2\UserManager;
 use Doctrine\ORM\EntityManagerInterface;
@@ -29,8 +29,7 @@ class UserController extends AbstractController
         TranslatorInterface $translator,
     ): Response {
         $user = $this->getUser();
-
-        $userForm = $this->createForm(UserSettingsType::class, $user)->handleRequest($request);
+        $userForm = $this->createForm(UserType::class, $user)->handleRequest($request);
 
         $passwordForm = $this->createForm(ChangePasswordFormType::class, null, [
             'checkCurrentPassword' => true,
