+++ src/ServiceV2/BeneficiaryClientLinkService.php
@@ -17,14 +17,14 @@ class BeneficiaryClientLinkService
     ) {
     }
 
-    public function linkBeneficiaryToClientWithName(Beneficiaire $beneficiary, string $clientName, string $externalId, ?string $memberExternalId = null): void
+    public function linkBeneficiaryToClientWithName(Beneficiaire $beneficiary, string $clientName, string $externalId, string $memberExternalId = null): void
     {
         if ($client = $this->clientRepository->findOneBy(['nom' => $clientName])) {
             $this->linkBeneficiaryToClient($beneficiary, $client, $externalId, $memberExternalId);
         }
     }
 
-    public function linkBeneficiaryToClient(Beneficiaire $beneficiary, Client $client, string $externalId, ?string $memberExternalId = null): void
+    public function linkBeneficiaryToClient(Beneficiaire $beneficiary, Client $client, string $externalId, string $memberExternalId = null): void
     {
         $beneficiary->addClientExternalLink($client, $externalId, $memberExternalId);
     }
