+++ templates/v2/user/_settings_form_block.html.twig
@@ -1,18 +1,8 @@
 {% if title %}
     <div class="col-12 mb-3">
         <div class="bg-primary text-white rounded py-1 mt-3">
-            <h5 class="text-center mb-0 ">{{ title }}</h5>
+            <h5 class="text-center mb-0 ">{{ title|trans }}</h5>
         </div>
     </div>
 {% endif %}
-<div class="col-12 col-md-6">
-    {{ include('v2/common/_field_list.html.twig', {fields: leftColumnFields}) }}
-</div>
-<div class="col-12 col-md-6">
-    {{ include('v2/common/_field_list.html.twig', {fields: rightColumnFields}) }}
-</div>
-{% if fullWidthColumnFields is defined %}
-    <div class="col-12">
-        {{ include('v2/common/_field_list.html.twig', {fields: fullWidthColumnFields}) }}
-    </div>
-{% endif %}
+{{ include('v2/common/_field_list.html.twig', {fields}) }}
